requirements.txt
fastapi
uvicorn[standard]
python-multipart
pdfplumber
PyMuPDF
httpx
python-dotenv
pydantic-settings
supabase
boto3
google-cloud-vision
Pillow
